download_service/requirements.txt
fastapi
uvicorn
pymongo

PyJWT
python-jose
requests
